removed inputfield component
