+++ client/src/components/EquationBox/EQBox.css
@@ -5,9 +5,10 @@
     background: #f7f7f700;
     font-family: 'Balsamiq Sans', cursive;
     z-index: 1;
-    display: inline-flexbox;
+    display: inline-block;
     margin: auto;
-    width: 20%;
+    width: 70%;
+    margin-top: 0;
     user-select: none;
     -moz-user-select: none;
     -khtml-user-select: none;
@@ -50,4 +51,12 @@
     -o-user-select: auto;
     pointer-events: auto;
     background: auto;
+}
+
+.answer {
+    text-align: center;
+    font-size: 80px;
+    width: 100px;
+    height: 140px;
+    float: left;
 }
