--- client/src/components/NumberField/NumberField.css
@@ -1,6 +1,0 @@
-.answer {
-    text-align: center;
-    font-size: 80px;
-    width: 100px;
-  
-}
